backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
alembic
databases[sqlite]
Jinja2
python-multipart
websockets
httpx
pandas
pandas-ta
PyYAML
loguru
pyotp
numpy

# Broker APIs
smartapi-python
logzero
websocket-client

# Testing
pytest
pytest-asyncio

# Notifications
